requirements.txt
flask
flask-login
flask-sqlalchemy
gunicorn
requests
python-telegram-bot